@@ -44,34 +44,33 @@
     "deploy:demo": "ng build zxing-scanner-demo --prod --base-href /ngx-scanner/ && npx ngh --dir=dist/zxing-scanner-demo"
   },
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/cdk": "~7.3.7",
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/material": "7.3.7",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
-    "@zxing/library": "^0.14.1",
+    "@angular/animations": "^8.0.2",
+    "@angular/cdk": "~8.0.1",
+    "@angular/common": "^8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/forms": "^8.0.2",
+    "@angular/material": "8.0.1",
+    "@angular/platform-browser": "^8.0.2",
+    "@angular/platform-browser-dynamic": "^8.0.2",
+    "@angular/router": "^8.0.2",
+    "@zxing/library": "^0.14.2",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.5.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.4",
-    "@angular-devkit/build-ng-packagr": "^0.12.3",
-    "@angular-devkit/core": "^7.0.4",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler": "^7.0.0",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular-devkit/core": "^8.0.4",
+    "@angular/cli": "~8.0.4",
+    "@angular/compiler": "^8.0.2",
+    "@angular/compiler-cli": "^8.0.2",
+    "@angular/language-service": "^8.0.2",
     "@compodoc/compodoc": "^1.1.1",
     "@types/jasmine": "^2.8.6",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^10.12.2",
     "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "^4.3.0",
+    "codelyzer": "^5.0.1",
     "core-js": "^2.5.5",
     "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "^4.2.1",
@@ -81,14 +80,15 @@
     "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^4.4.0",
+    "ng-packagr": "^5.1.0",
     "npm-install-peers": "^1.2.1",
     "protractor": "^5.3.1",
     "ts-node": "^7.0.1",
-    "tsickle": "^0.33.1",
-    "tslint": "^5.9.1",
-    "typescript": "~3.1.3",
-    "zone.js": "^0.8.26"
+    "tsickle": "^0.35.0",
+    "tslib": "^1.10.0",
+    "tslint": "^5.18.0",
+    "typescript": "~3.4.5",
+    "zone.js": "~0.9.1"
   },
   "collective": {
     "type": "opencollective",
